+++ src/tespy/connections/__init__.py
@@ -2,4 +2,4 @@
 
 from .bus import Bus  # noqa: F401
 from .connection import Connection  # noqa: F401
-from .ref import Ref  # noqa: F401
+from .connection import Ref  # noqa: F401
